// server/interface/request.interface.ts
export interface SendSmsParams {
  captcha?: string
  captchaId?: string
  mobile: string
}
// ...
export interface UpdatePhoneSendMessageParams {
  oldPhone: string
  captcha: string
  captchaId: string
  type: string
  newPhone?: string
}
export interface VerificationUpdatePhoneMessageParams {
  smsCode: string
  newPhone: string
}

export interface VerificationMessageParmas {
  smsCode: string
}